\name{[.mutPPAs}
\alias{[.mutPPAs}
\title{Subset method for "mutPPAs" objects}
\usage{
  \method{[}{mutPPAs} (x, i)
}
\arguments{
  \item{x}{mutPPAs object}

  \item{i}{a scalar indicating the nucleotide site to
  summarise}

  \item{...}{ignored}
}
\description{
  Subset method for "mutPPAs" objects
}

